Update constant.ts
I found that two different constants have the same value: contextMenuClick and columnMenuClick. This is the reason for a bug when the user tries to work with the click on the column menu item event emitter. I just edited columnMenuClick constant and set the value to 'columnMenuClick' as it should be.